Make file check asynchronous
The CSV factory method `fromPath` performs I/O, which means that it should return a Promise, as well as its caller, `FileSystemEmployees.fromCsv`.  This design makes the factory methods asynchronous, so that when they complete, the objects are ready to run synchronously.

This approach would not work well if the CSV file was long enough that it was slow/expensive to load it all into memory first.  In that case, moving the asynchronous part to `all` by streaming the read rows might be more effective.