Rename fs module import
The previous changes used an expand-and-contract approach to add in the asynchronous `fs` module version without requiring us to change all the file system operations to asynchronous at once.  While in practice that would have been easy enough to do in one commit, I wanted to see how easily it would be to take an expand/contract approach.  (It turned out it was possible and not all that hard, since we could import the two modules side by side)